code4C/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.4.1
contourpy==1.1.0
cycler==0.11.0
dlib==19.24.2
filelock==3.12.3
fonttools==4.42.1
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib @ file:///croot/joblib_1685113087166/work
kiwisolver==1.4.5
lit==17.0.0rc4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1651020413938/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-contrib-python==4.4.0.44
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn @ file:///croot/scikit-learn_1690978916802/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1653073867187/work
six==1.16.0
sympy==1.12
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
torch==2.0.1
torchaudio==2.0.2
torchstat==0.0.7
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
zipp==3.16.2
